LDA transition to Py3,new regex,support for 2017-18
The code is now compatible with Python 3. A new pre-tested (files associated with the final test added) regex has been defined to identify posts related to marijuana's legal status on Reddit. The parser can now parse posts until the end of 2018, support for which was previously lacking because of different compression algorithms in the source dataset.
Note that the neural network code is not finalized and only the LDA code has been debugged.